+++ packages/fresh_http/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.1
+
+- fix: new requests during in-flight token refresh no longer sent with stale token ([#141](https://github.com/felangel/fresh/issues/141), [#142](https://github.com/felangel/fresh/pull/142))
+
 # 0.1.0
 
 - feat: initial implementation that has feature parity with `fresh_dio` and `fresh_graphql`